feat: recruitment api
refs #13

+++ src/main/java/com/jeari/repository/RecruitmentRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface RecruitmentRepository extends JpaRepository<Recruitment, Long> {
+public interface RecruitmentRepository extends JpaRepository<Recruitment, Integer> {
 
     // 클럽 ID로 최신 모집 공고 1개 조회 (startDate 기준 내림차순)
     Optional<Recruitment> findFirstByClubIdOrderByStartDateDesc(Integer clubId);
